feat: add auto-reconnect & allow connections to be reused
Prior to this commit, SurrealWebSocketConnection could not `connect -> disconnect -> connection`. An exception would be thrown on the last connect call. This has been fixed by creating a new WebSocket on every disconnect. This is in line with how the WebSocket library is intended to be used.

Along with making SurrealWebSocketConnection reusable, I've also implemented auto-reconnect functionality.