Added the JDBC Tests to the build.gradle to use the env vars
Started to implement some Parts of the connection and Driver to be able to connect initially.

Copied the TestUtils to the jdbc tests and added 2 dev-wise params

Added the exception package for jdbc, could come in handy in the future

Implemented the connect method of the driver, getObject of the Statement and some other and the query method.

Added a few tests for the SurrealJDBCDriverTest and SurrealJDBCStatementTest

Still WIP.